worker/requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.0

# LangChain + LangGraph
langchain==0.3.14
langchain-google-genai==2.0.8
langgraph==0.2.60

# Database
psycopg2-binary==2.9.10
pgvector==0.3.6
sqlalchemy==2.0.36

# Google APIs
google-auth==2.37.0
google-api-python-client==2.159.0

# Utilities
httpx==0.28.1
python-dotenv==1.0.1
cryptography==44.0.0
structlog==24.4.0
tenacity==9.0.0
